Allow disabling debugging messages.
Add test for LWASM_NODEBUG symbol to disable compiling the actual debug
message handling code.

Also, prevent building the debug message arguments if the debug level is not
going to display the message anyway. This comes at the possible expense of
larger code due to wrapping the debug_mesage() function in a macro that
tests the debug level directly. This should prevent calling expensive things
like building a dump of an expression when it is not required.