Allow FCB-like syntax after constant in FCC under m80ext
This is an abomination (say what you really mean!) but for compatibilty with
Macro-80C, allow handling of FCB style syntax after a string constant
specified for FCC. Mostly, there are better ways to do this in new code so
it is only accepted under the m80ext pragma.

Thanks to Erik G <erik@6809.org> for the patch.