Add option to specify tab width in listing
There is much insistence that generating the asm listing should absolutely
not ever in the entire expanse of space and time expand tabs. lwasm,
however, does so, and it does so for a reason. That reason is that there are
enough users that have enough trouble even understanding what a tab is that
it was simpler to just expand them.

That said, having a means to actually specify the tab width is pretty much
required if you're going to do that, and having a way to disable such
expansion is also a good idea.

Thanks to Erik G <erik@6809.org> for the base patch to handle this.