Add << prefix to force 5 bit offsets in indexed modes
Rounding out the compliment of operand size prefixes, we now have "<<" to
mean "force 5 bits".  According to Steve Bjork, this was the "official" way
to do this since 1980.  However, I have no official Motorola source for
that.  It does suggest that the choice of "<<" is consistent with other
(historical) assemblers, though.

Either way, it seems the most logical choice while avoiding any conflicts
with legal source code, so "<<" it is.