Don't rely on undefined memory before start of instab
When encountering a symbol on a line by itself, cl->insn was used to index
into the instab array. Unfortunately, in that case, cl->insn will be -1
which will refer to memory before the start of instab. Depending on the
compiler and linker, that could be anything.

This would only have a visible effect on the OS9 target which has separate
data and code address counters. On other targets, the two counters are kept
in sync.

This patch should short circuit that unfortunate code issue.