Update raw output to work with RMB only definitions at the start
It seems useful to be able to define a number of symbols at the start of the
source file without generating a run of NUL output. This update adjusts the
output to ignore any number of statements that generate no output as long as
there is a subsequent ORG statement. If there is no ORG statement after
them, the NUL sequence should still appear.