Allow MOD directive to accept only the standard base header parameters
There are cases where only the standard OS9 module header fields are to be
provided. This could be, for instance, in the case of a nonstandard user
type module. So allow the execution offset and data size fields to be
omitted entirely and handle the module length correctly as well.