Initial implementation of PHASE/DEPHASE
Create new psuedo ops PHASE and DEPHASE which allow specifying the address
for the assembler to assume for the code between the PHASE and DEPHASE pair.
So if you have code that is ORG'd to 0x8000 but you have a block of code
that will be copied to another address at runtime and executed there, say
0x4000, you can specify PHASE 0x4000 before that block of code and DEPHASE
after it and the assembler will generate the code there as though there had
been an ORG 0x4000 *except* it will continue outputting the code at the same
address it would have had without the PHASE statement. After DEPHASE, the
current assembly address resumes its natural value.