Add :: operaetor for pasting two LSBs into a single word
Based on an idea pioneered by spc476 on the CoCo Discord server. This
allows one to avoid a lot of (expr)*256+(expr) constructions. In fact,
that is exactly what the :: operator does except if adds an implied
"& 0xff" on each operand.

This operator *will not* appear in generated object files. It's solely
present as syntactic sugar and actually converts to other operators
properly grouped in the resulting expression.